+++ includes/class-wc-gateway-admin-form-fields.php
@@ -387,7 +387,7 @@ class WC_Gateway_Admin_Form_Fields {
 				'title'       => __( 'Payment group 2 methods', 'wc-maksuturva' ),
 				'desc_tip'    => true,
 				'description' => __( 'Collated payment methods, group 2 methods', 'wc-maksuturva' ),
-				'default'     => get_option( 'collated_group2_methods', 'FI50,FI51,FI52,FI53,FI54,PIVO,SIIR' ),
+				'default'     => get_option( 'collated_group2_methods', 'FI50,FI51,FI52,FI53,FI54,SIIR' ),
 			),
 			'collated_group3_title'                       => array(
 				'type'        => 'textfield',