+++ includes/class-wc-payment-method-select.php
@@ -238,8 +238,7 @@ class WC_Payment_Method_Select {
 			}
 
 			foreach ( $available_payment_methods['paymentmethod'] as $key => $payment_method ) {
-				if ( in_array( substr( $payment_method['code'], 0, 3 ), array( 'FI5' ) ) || in_array( substr( $payment_method['code'], 0, 4 ), array( 'PIVO' ) )
-					|| in_array( substr( $payment_method['code'], 0, 4 ), array( 'SIIR' ) ) ) {
+				if ( in_array( substr( $payment_method['code'], 0, 3 ), array( 'FI5' ) ) || in_array( substr( $payment_method['code'], 0, 4 ), array( 'SIIR' ) ) ) {
 					$payment_type_payment_methods['credit-card-and-mobile'][] = $payment_method;
 					unset( $available_payment_methods['paymentmethod'][ $key ] );
 				}